extract render filter logic into util
